Finished PairsPMI

+++ src/main/java/ca/uwaterloo/cs451/a1/PairsPMI.java
@@ -118,7 +118,7 @@ public class PairsPMI extends Configured implements Tool {
         @Override
         public void setup(Context context) throws IOException {
             List<String> lines;
-            numberOfLines = context.getConfiguration().getInt("numberOfLines", -1);
+            numberOfLines = context.getConfiguration().getDouble("numberOfLines", -1);
             System.out.println("number of lines in file : " + numberOfLines);
             try {
                 lines = Files.readLines(new File(context.getCacheFiles()[0].toString()), StandardCharsets.UTF_8);
